+++ src/cli/global.rs
@@ -7,7 +7,8 @@ use sysinfo::{MemoryRefreshKind, RefreshKind, System};
 use tracing::Level;
 
 use crate::cli::commands::{
-    github::GitHubArgs, gitlab::GitLabArgs, rules::RulesArgs, scan::ScanArgs,
+    github::GitHubArgs, gitlab::GitLabArgs, precommit::PrecommitArgs, rules::RulesArgs,
+    scan::ScanArgs,
 };
 
 #[deny(missing_docs)]
@@ -62,6 +63,10 @@ pub enum Command {
     /// Manage rules
     #[command(alias = "rule")]
     Rules(RulesArgs),
+
+    /// Manage Kingfisher as a Git pre-commit hook
+    #[command(name = "precommit")]
+    Precommit(PrecommitArgs),
 }
 
 pub static RAM_GB: Lazy<Option<f64>> = Lazy::new(|| {